Add option parsing to inventory.py.
The aim of this PR is to add option parsing to inventory.py.
In particular, it adds --no-app-nodes option, which excludes
app nodes from the [nodes] group.  Suppressing the app nodes
is useful during cluster scaleups, as it reduces ansible tasks
on nodes which are already part of the cluster.  As a result,
cluster scaleup with ~100s of nodes are much faster.